# docs/swagger.yaml
definitions:
...
  request_model.CreateCareerDepartmentRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  request_model.ReviewApprovalRequest:
    properties:
      rejected_reason:
        type: string
    type: object
...
  request_model.UpdateCareerDepartmentRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  response.Base:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      validation:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /career-departments:
    get:
      consumes:
      - application/json
      description: Get all career departments with pagination (Admin & Superadmin)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get All Career Departments (CMS)
      tags:
      - Career Department
    post:
      consumes:
      - application/json
      description: Admin creates a new Career Department approval request
      parameters:
      - description: Create Career Department Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.CreateCareerDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Create Career Department (Admin)
      tags:
      - Career Department
  /career-departments/{id}:
    delete:
      description: Admin directly deletes a Career Department (no approval needed)
      parameters:
      - description: Career Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Delete Career Department (Admin)
      tags:
      - Career Department
    get:
      consumes:
      - application/json
      description: Get career department detail by ID (Admin & Superadmin)
      parameters:
      - description: Career Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get Career Department By ID (CMS)
      tags:
      - Career Department
    put:
      consumes:
      - application/json
      description: Admin creates a Career Department update approval request
      parameters:
      - description: Career Department ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Career Department Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.UpdateCareerDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Update Career Department (Admin)
      tags:
      - Career Department
  /career-departments/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Admin uploads an image for a career department to S3
      parameters:
      - description: Career Department ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file (JPEG, PNG, WebP)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Upload Career Department Image (Admin)
      tags:
      - Career Department
  /career-departments/approvals:
    get:
      consumes:
      - application/json
      description: Get all career department approval requests with pagination (Superadmin)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Get Career Department Approval Requests
      tags:
      - Career Department Approvals
  /career-departments/approvals/{id}/approve:
    post:
      consumes:
      - application/json
      description: Superadmin approves a career department approval request
      parameters:
      - description: Approval Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Approve Career Department Approval Request
      tags:
      - Career Department Approvals
  /career-departments/approvals/{id}/reject:
    post:
      consumes:
      - application/json
      description: Superadmin rejects a career department approval request
      parameters:
      - description: Approval Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request_model.ReviewApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Reject Career Department Approval Request
      tags:
      - Career Department Approvals
  /career-departments/approvals/{id}/takedown:
    post:
      consumes:
      - application/json
      description: Superadmin takes down a career department and sets status to Draft
      parameters:
      - description: Approval Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      security:
      - BearerAuth: []
      summary: Takedown Career Department Approval Request
      tags:
      - Career Department Approvals
...
  /public/career-departments:
    get:
      consumes:
      - application/json
      description: Get all approved career departments (no authentication required)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      summary: Get Public Career Departments
      tags:
      - Career Department Public
  /public/career-departments/{id}:
    get:
      consumes:
      - application/json
      description: Get a single approved career department by ID (no authentication
        required)
      parameters:
      - description: Career Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      summary: Get Public Career Department By ID
      tags:
      - Career Department Public
...
securityDefinitions:
  BearerAuth:
    description: 'Insert JWT Token with format: **Bearer {your token}**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"